requirements.txt
Django
celery
dotenv
django-cors-headers
django-countries
graphene-django
django-graphql-auth
django-graphql-jwt==0.3.0 
PyJWT==1.7.1
djangorestframework
django-allauth
psycopg2-binary
PyGithub
gunicorn
beautifulsoup4
cryptography
django-celery-beat
django-celery-results
redis
python-dotenv